requirements.txt
absl-py==1.0.0
ase==3.22.1
astunparse==1.6.3
black==22.1.0
cachetools==4.2.4
catboost==1.0.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.4
cycler==0.11.0
flatbuffers==2.0
fonttools==4.28.5
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19.1
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.0
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lightgbm==3.3.2
Markdown==3.3.6
matplotlib==3.5.1
megnet==1.2.9
monty==2021.12.1
mpmath==1.2.1
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
palettable==3.3.0
pandas==1.3.5
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.5.1
plotly==5.5.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymatgen==2022.0.17
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
spglib==1.16.3
sympy==1.9
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.0.0
tomli==2.0.0
torch==1.10.2
tqdm==4.62.3
typing-extensions==4.0.1
uncertainties==3.1.6
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
xgboost==1.5.2
zipp==3.6.0
